+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - By default, run only 100 optimization steps in `build_random_cell`.
+- Wrap atoms back into the cell when writing PDB trajectory files, for nicer visual.
+
 
 ## [0.0.0] - 2024-10-06